+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed return type of `getDefaultCollectionFields` and custom collection `fields`
 - Renamed `typesenseOptions` to `serverConfig` in PluginOptions and add doc comments.
 - Added VN locale.
+- Added `indexCodeBlocks` plugin config option to control whether code blocks should be indexed or not. Defaults to `false` to avoid indexing large code blocks which may not be relevant for search results and can bloat the index.
 
 ## 0.0.1